move standard pipeline eval tests to separate test file
